+++ src/main/java/com/pos/hyper/repository/ProductRepository.java
@@ -1,7 +1,7 @@
 package com.pos.hyper.repository;
 
-import com.pos.hyper.model.Product;
+import com.pos.hyper.model.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
